TOOLKIT-1887 (Problem with deployment package generation)

+++ src/plugins/org.tigris.mtoolkit.dpeditor/src/org/tigris/mtoolkit/dpeditor/installation/DPPFileProvider.java
@@ -35,15 +35,15 @@ import org.tigris.mtoolkit.dpeditor.util.DPPUtil;
 import org.tigris.mtoolkit.util.DPPFile;
 
 public final class DPPFileProvider extends WorkspaceFileProvider {
-
+  
   public class DPPFileItem extends WorkspaceFileItem {
     public File dpFile;
     private boolean delete = false;
-
+    
     public DPPFileItem(IFile file, String mimeType) {
       super(file, mimeType);
     }
-
+    
     public void dispose() {
       super.dispose();
       if (delete) {
@@ -55,14 +55,14 @@ public final class DPPFileProvider extends WorkspaceFileProvider {
         }
       }
     }
-
+    
     public InputStream getInputStream() throws IOException {
       if (preparedFile != null) {
         return new FileInputStream(preparedFile);
       }
       return new FileInputStream(dpFile);
     }
-
+    
     public String getLocation() {
       if (preparedFile != null) {
         return preparedFile.getAbsolutePath();
@@ -72,15 +72,16 @@ public final class DPPFileProvider extends WorkspaceFileProvider {
       }
       throw new IllegalStateException("Installation item wasn't initialized correctly, missing location to base file");
     }
-
+    
     public void setLocation(File file) {
       if (preparedFile != null && !preparedFile.equals(file)) {
         // delete the previous prepared item
-        preparedFile.delete();
+        if (delete)
+          preparedFile.delete();
       }
       preparedFile = file;
     }
-
+    
     public IStatus prepare(IProgressMonitor monitor, Map properties) {
       try {
         Display dis = PlatformUI.getWorkbench().getDisplay();
@@ -93,7 +94,10 @@ public final class DPPFileProvider extends WorkspaceFileProvider {
         dpFile = new File(dppFile.getBuildInfo().getBuildLocation() + "/" + dppFile.getBuildInfo().getDpFileName());
         delete = !dpFile.exists();
         preparedFile = (delete) ? dpFile : null;
-        DPPUtil.generateDeploymentPackage(dppFile, monitor, file.getProject(), DPPUtil.TYPE_QUICK_BUILD_DPP);
+        IStatus status = DPPUtil.generateDeploymentPackage(dppFile, monitor, file.getProject(), DPPUtil.TYPE_QUICK_BUILD_DPP);
+        if (!status.isOK()) {
+          return status;
+        }
         if (properties != null && "Dalvik".equalsIgnoreCase((String) properties.get("jvm.name"))
             && !AndroidUtils.isDpConvertedToDex(dpFile)) {
           File convertedFile = new File(DPActivator.getDefault().getStateLocation() + "/dex/" + dpFile.getName());
@@ -109,25 +113,29 @@ public final class DPPFileProvider extends WorkspaceFileProvider {
         return new Status(Status.ERROR, DPActivator.PLUGIN_ID, "Failed to prepare file for installation");
       }
     }
-
+    
   } // end of DPPFileItem
-
+  
   public IStatus prepareItems(List items, Map properties, IProgressMonitor monitor) {
     if (items != null) {
+      IStatus status;
       for (int i = 0; i < items.size(); i++) {
         Object item = items.get(i);
         if (item instanceof DPPFileItem) {
-          ((DPPFileItem) item).prepare(monitor, properties);
+          status = ((DPPFileItem) item).prepare(monitor, properties);
+          if (!status.isOK()) {
+            return status;
+          }
         }
       }
     }
     return Status.OK_STATUS;
   }
-
+  
   public String getName() {
     return "Deployment packages provider";
   }
-
+  
   public InstallationItem getInstallationItem(Object resource) {
     return new DPPFileItem(getFileFromGeneric(resource), mimeType);
   }
